+++ app/src/main/java/com/amsterdam/cutetudee/presentation/screens/home/HomeScreen.kt
@@ -34,9 +34,9 @@ import com.amsterdam.cutetudee.presentation.component.chip.tast_status.TaskStatu
 import com.amsterdam.cutetudee.presentation.component.custom_padding.bottomNavigationBarPadding
 import com.amsterdam.cutetudee.presentation.component.custom_snack_bar.CustomSnackBarStatus
 import com.amsterdam.cutetudee.presentation.navigation.Screen
-import com.amsterdam.cutetudee.presentation.component.sharedComponent.OverlayBoxContent
-import com.amsterdam.cutetudee.presentation.component.sharedComponent.TaskSection
-import com.amsterdam.cutetudee.presentation.component.sharedComponent.TopCuteTudeeAppBar
+import com.amsterdam.cutetudee.presentation.screens.home.sharedComponent.OverlayBoxContent
+import com.amsterdam.cutetudee.presentation.screens.home.sharedComponent.TaskSection
+import com.amsterdam.cutetudee.presentation.screens.home.sharedComponent.TopCuteTudeeAppBar
 import com.amsterdam.cutetudee.presentation.screens.common.AddEditTaskInteractionListener
 import com.amsterdam.cutetudee.presentation.screens.common.AddEditTaskUiState
 import com.amsterdam.cutetudee.presentation.component.AddOrEditTaskBottomSheet
